+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexLanguageIsNotSupportedInspection.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -47,9 +47,10 @@ class ImpexLanguageIsNotSupportedInspection : LocalInspectionTool() {
             }
                 .trim()
 
-            val supportedLanguages = PropertiesService.getInstance(psi.project).getLanguages()
+            val propertiesService = PropertiesService.getInstance(psi.project)
+            val supportedLanguages = propertiesService.getLanguages()
 
-            if (supportedLanguages.contains(language)) return
+            if (propertiesService.containsLanguage(language, supportedLanguages)) return
 
             holder.registerProblem(
                 psi,