+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/lang/annotation/PolyglotQueryAnnotator.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -49,8 +49,13 @@ class PolyglotQueryAnnotator : Annotator {
                 BIND_PARAMETER -> highlight(BIND_PARAMETER, holder, element)
                 LOCALIZED_NAME -> {
                     val language = element.text.trim()
-                    val supportedLanguages = PropertiesService.getInstance(element.project).getLanguages()
-                    if (!supportedLanguages.contains(language)) {
+
+                    val propertiesService = PropertiesService.getInstance(element.project)
+                    val supportedLanguages = propertiesService.getLanguages()
+
+                    if (propertiesService.containsLanguage(language, supportedLanguages)) {
+                        highlight(LOCALIZED_NAME, holder, element)
+                    } else {
                         highlightError(
                             holder, element,
                             message(
@@ -59,8 +64,6 @@ class PolyglotQueryAnnotator : Annotator {
                                 supportedLanguages.joinToString()
                             )
                         )
-                    } else {
-                        highlight(LOCALIZED_NAME, holder, element)
                     }
                 }
             }