+++ src/com/intellij/idea/plugin/hybris/properties/PropertiesService.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,6 +23,8 @@ import com.intellij.openapi.project.Project
 interface PropertiesService {
 
     fun getLanguages(): Set<String>
+
+    fun containsLanguage(language: String, supportedLanguages: Set<String>): Boolean
 
     companion object {
         fun getInstance(project: Project): PropertiesService = project.getService(PropertiesService::class.java)